+++ instrumentation/opentelemetry-instrumentation-system-metrics/tests/test_system_metrics.py
@@ -136,6 +136,9 @@ class TestSystemMetrics(TestBase):
             observer_names.append(
                 f"process.runtime.{self.implementation}.gc_count",
             )
+            observer_names.append(
+                "cpython.gc.collections",
+            )
         if sys.platform != "darwin":
             observer_names.append("system.network.connections")
 
@@ -946,13 +949,38 @@ class TestSystemMetrics(TestBase):
     def test_runtime_get_count(self, mock_gc_get_count):
         mock_gc_get_count.configure_mock(**{"return_value": (1, 2, 3)})
 
-        expected = [
+        expected_gc_count = [
             _SystemMetricsResult({"count": "0"}, 1),
             _SystemMetricsResult({"count": "1"}, 2),
             _SystemMetricsResult({"count": "2"}, 3),
         ]
         self._test_metrics(
-            f"process.runtime.{self.implementation}.gc_count", expected
+            f"process.runtime.{self.implementation}.gc_count",
+            expected_gc_count,
+        )
+
+    @mock.patch("gc.get_stats")
+    @skipIf(
+        python_implementation().lower() == "pypy", "not supported for pypy"
+    )
+    def test_runtime_get_gc_collections(self, mock_gc_get_stats):
+        mock_gc_get_stats.configure_mock(
+            **{
+                "return_value": [
+                    {"collections": 10, "collected": 100, "uncollectable": 1},
+                    {"collections": 20, "collected": 200, "uncollectable": 2},
+                    {"collections": 30, "collected": 300, "uncollectable": 3},
+                ]
+            }
+        )
+        expected_gc_collections = [
+            _SystemMetricsResult({"generation": "0"}, 10),
+            _SystemMetricsResult({"generation": "1"}, 20),
+            _SystemMetricsResult({"generation": "2"}, 30),
+        ]
+        self._test_metrics(
+            "cpython.gc.collections",
+            expected_gc_collections,
         )
 
     @mock.patch("psutil.Process.num_ctx_switches")